+++ apps/auth/src/application/users/repositories/events/constants.ts
@@ -1,0 +1,3 @@
+export const USERS_TOPICS = {
+  USER_CREATED: 'user-created',
+};
