+++ apps/auth/src/application/users/repositories/events/implementation/in-memory-users-events.repository.ts
@@ -1,7 +1,9 @@
 // users.repository.ts
 import { Injectable } from '@nestjs/common';
-import { User } from '../../entities/user.entity';
-import { UsersEventsRepository } from './users-events.repository';
+import { User } from '../../../entities/user.entity';
+import { USERS_TOPICS } from '../constants';
+import { IUserCreatedMessagePayload } from '../types';
+import { UsersEventsRepository } from '../users-events.repository';
 
 export interface IHeaders {
   [key: string]: Buffer | string | (Buffer | string)[] | undefined;
@@ -18,16 +20,16 @@ export class InMemoryUsersEventsRepository implements UsersEventsRepository {
   private usersCreatedTopic: CreatedUserMessage[] = [];
 
   async publishUserCreated(user: User): Promise<void> {
+    const messagePayload: IUserCreatedMessagePayload = {
+      email: user.email,
+      id: user.id,
+    };
+
     this.usersCreatedTopic.push({
       key: user.email,
-      value: JSON.stringify({
-        id: user.id,
-        email: user.email,
-      }),
-      headers: {
-        role: 'default',
-      },
+      value: JSON.stringify(messagePayload),
     });
-    console.log('user-created', this.usersCreatedTopic);
+
+    console.log(USERS_TOPICS['USER_CREATED'], this.usersCreatedTopic);
   }
 }
