+++ apps/auth/src/application/users/repositories/events/implementation/kafka-users-events.repository.ts
@@ -1,9 +1,11 @@
 // users.repository.ts
 import { Injectable } from '@nestjs/common';
 import { Message } from 'kafkajs';
-import { KafkaService } from '../../../../infra/events/kafka/kafka.service';
-import { User } from '../../entities/user.entity';
-import { UsersEventsRepository } from './users-events.repository';
+import { KafkaService } from '../../../../../infra/events/kafka/kafka.service';
+import { User } from '../../../entities/user.entity';
+import { USERS_TOPICS } from '../constants';
+import { IUserCreatedMessagePayload } from '../types';
+import { UsersEventsRepository } from '../users-events.repository';
 
 @Injectable()
 export class KafkaUsersEventsRepository implements UsersEventsRepository {
@@ -12,19 +14,18 @@ export class KafkaUsersEventsRepository implements UsersEventsRepository {
   async publishUserCreated(user: User): Promise<void> {
     const producer = await this.kafkaService.createProducer();
 
+    const messagePayload: IUserCreatedMessagePayload = {
+      email: user.email,
+      id: user.id,
+    };
+
     const message: Message = {
       key: user.email,
-      value: JSON.stringify({
-        email: user.email,
-        id: user.id,
-      }),
-      headers: {
-        role: 'default',
-      },
+      value: JSON.stringify(messagePayload),
     };
 
     const result = await producer.send({
-      topic: 'user-created',
+      topic: USERS_TOPICS['USER_CREATED'],
       messages: [message],
     });
