+++ apps/auth/src/application/users/repositories/events/types.ts
@@ -1,0 +1,4 @@
+export interface IUserCreatedMessagePayload {
+  id: string;
+  email: string;
+}
