+++ apps/auth/src/application/users/repositories/storage/implementation/index.repository.ts
@@ -1,0 +1,16 @@
+import { Logger } from '@nestjs/common';
+import { InMemoryUsersStorageRepository } from './in-memory-users-storage.repository';
+import { PrismaUsersStorageRepository } from './prisma-users-storage.repository';
+
+const isPersistentStorageEnabled =
+  process.env['PERSISTENT_STORAGE_ENABLED'] === 'true';
+
+export const UsersStorageRepositoryImplementation = isPersistentStorageEnabled
+  ? PrismaUsersStorageRepository
+  : InMemoryUsersStorageRepository;
+
+Logger.log(
+  isPersistentStorageEnabled
+    ? 'Using persistent storage...'
+    : 'Using in memory storage...'
+);
